requirements.txt
Flask
python-dotenv
streamlit
requests
Flask-Cors
gunicorn
groq
langchain
langchain-groq
langchain-core